requirements.txt
glcontext==3.0.0
glfw==2.9.0
moderngl==5.12.0
networkx==3.4.2
numpy==2.2.6
pillow==11.2.1
pygame==2.6.1
pyglm==2.8.2
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
